more windows search and group policies disabled

+++ GameModeWin/ConfigServices.cs
@@ -178,33 +178,67 @@ namespace GameModeWin
             servicePath1.SetValue("Start", 4);
             servicePath1.Close();
 
+            disableTelemtry();
+            appCompatOff();  
+        }
+
+        public void appCompatOff()
+        {
+            RegistryKey servicePath3 = Registry.CurrentUser.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
+
+            if (servicePath3.OpenSubKey("AppCompat") == null)
+            {
+                servicePath3.Close();
+                RegistryKey sp1 = Registry.CurrentUser.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
+
+                sp1 = sp1.CreateSubKey("AppCompat");
+                sp1.Close();
+                servicePath3 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
+            }
+            servicePath3 = servicePath3.OpenSubKey("AppCompat", true);
+
+            servicePath3.SetValue("DisablePCA", 1);
+            servicePath3.Close();
+        }
+
+        
+        public void disableTelemtry()
+        {
             RegistryKey servicePath3 = Registry.CurrentUser.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
 
             if (servicePath3.OpenSubKey("DataCollection") == null)
             {
                 servicePath3.Close();
                 RegistryKey sp1 = Registry.CurrentUser.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
-               
+
                 sp1 = sp1.CreateSubKey("DataCollection");
                 sp1.Close();
-                servicePath3 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true); 
+                servicePath3 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
             }
             servicePath3 = servicePath3.OpenSubKey("DataCollection", true);
 
             /// they say that it will be equal to 1 if it's set to 0, but why they got 0?
             /// this can be commented and maybe the result will be the same
             /// but if it will be not set, then collected data will be set through settings manually
+            /// disable this value here and commentout same deletevalue in notifyOn
             servicePath3.SetValue("AllowTelemetry", 0);
             servicePath3.Close();
-              
+
         }
+
+
 
         public void notifyOn()
         {
             RegistryKey servicePath = Registry.CurrentUser.OpenSubKey("Software\\Policies\\Microsoft\\Windows\\CurrentVersion\\PushNotifications", true);
-
-            servicePath.DeleteValue("NoToastApplicationNotification");
-            servicePath.DeleteValue("NoTileApplicationNotification");
+            if (servicePath.GetValue("NoToastApplicationNotification") != null)
+            {
+                servicePath.DeleteValue("NoToastApplicationNotification");
+            }
+            if (servicePath.GetValue("NoTileApplicationNotification") != null)
+            {
+                servicePath.DeleteValue("NoTileApplicationNotification");
+            }
             servicePath.Close();
 
             RegistryKey servicePath1 = Registry.LocalMachine.OpenSubKey("System\\CurrentControlSet\\Services\\DiagTrack", true);
@@ -213,8 +247,18 @@ namespace GameModeWin
             servicePath1.Close();
 
             RegistryKey servicePath3 = Registry.CurrentUser.OpenSubKey("Software\\Policies\\Microsoft\\Windows\\DataCollection", true);
-            servicePath3.DeleteValue("AllowTelemetry");
+            if (servicePath3.GetValue("AllowTelemetry") != null)
+            {
+                servicePath3.DeleteValue("AllowTelemetry");
+            }
             servicePath3.Close();
+
+            RegistryKey servicePath4 = Registry.CurrentUser.OpenSubKey("Software\\Policies\\Microsoft\\Windows\\AppCompat", true);
+            if (servicePath4.GetValue("DisablePCA") != null)
+            {
+                servicePath4.DeleteValue("DisablePCA");
+            }
+            servicePath4.Close();
 
         }
 
@@ -242,8 +286,38 @@ namespace GameModeWin
                 sp2.Close();
             }
 
+            disableWsearch();
 
 
+        }
+
+        public void disableWsearch()
+        {
+            RegistryKey sp3 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
+            if (sp3.OpenSubKey("Windows Search") == null)
+            {
+
+                RegistryKey sp4 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows", true);
+                sp4.CreateSubKey("Windows Search", 0);
+                sp4.Close();
+            }
+            sp3.Close();
+            RegistryKey sp5 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows\\Windows Search", true);
+
+            sp5.SetValue("AllowCortana", 0);
+            sp5.SetValue("DisableBackoff", 1);
+
+            if (sp5.OpenSubKey("PreventIndexingCertainPaths") == null)
+            {
+                RegistryKey sp6 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows\\Windows Search", true);
+                sp6.CreateSubKey("PreventIndexingCertainPaths", 0);
+                sp6.Close();
+            }
+            sp5.Close();
+
+            RegistryKey sp7 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows\\Windows Search\\PreventIndexingCertainPaths", true);
+            sp7.SetValue("C:\\", "C:\\");
+            sp7.SetValue("D:\\", "D:\\");
 
 
         }
@@ -268,8 +342,26 @@ namespace GameModeWin
                 sp2.Close();
             }
 
+            enableWSearch();
 
+        }
 
+        public void enableWSearch()
+        {
+            RegistryKey sp5 = Registry.LocalMachine.OpenSubKey("Software\\Policies\\Microsoft\\Windows\\Windows Search", true);
+            if (sp5.GetValue("AllowCortana") != null)
+            {
+                sp5.DeleteValue("AllowCortana");
+            }
+            if (sp5.GetValue("DisableBackoff") != null)
+            {
+                sp5.DeleteValue("DisableBackoff");
+            }
+            if (sp5.GetValue("PreventIndexingCertainPaths") != null)
+            {
+                sp5.DeleteSubKey("PreventIndexingCertainPaths");
+            }
+            sp5.Close();
         }
